% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gotm_sp.R
\name{run_gotm_sp}
\alias{run_gotm_sp}
\title{Run the GOTM-Selmaprotbas model}
\usage{
run_gotm_sp(
  sim_folder = ".",
  yaml_file = "gotm.yaml",
  verbose = FALSE,
  args = character()
)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{yaml_file}{filepath; to file with GOTM setup. Defaults to "gotm.yaml"}

\item{verbose}{Save output as character vector. Defaults to FALSE}

\item{args}{Optional arguments to pass to GOTM executable}
}
\description{
This runs the GOTM-Selmaprotbas model on the specific simulation stored in \code{sim_folder}.
}
\examples{
sim_folder <- system.file("extdata", package = "SelmaprotbasR")
run_gotm_sp(sim_folder)
}
\author{
Tadhg Moore, Jorrit Mesman
}
\keyword{methods}
